backend/requirements.txt
livekit-agents[openai,silero]~=1.1.5
livekit-plugins-silero~=1.1.5
livekit-plugins-openai~=1.1.5
python-dotenv
requests
aiohttp~=3.10
